CandSAPI/requirements.txt
asgiref==3.7.2
Django==3.2.24
django-cors-headers==4.1.0
django-filter==23.5
djangorestframework==3.14.0
importlib-metadata==6.7.0
Markdown==3.4.4
pytz==2024.1
sqlparse==0.4.4
typing-extensions==4.7.1
zipp==3.15.0
